+++ examples/use-cases/cloud_router_aws/variables.tf
@@ -95,11 +95,6 @@ variable "pf_cr_asn" {
   default  = 4556 # PacketFabric ASN
   nullable = false
 }
-# Parameter deprecated
-variable "pf_cr_scope" {
-  type    = string
-  default = "private"
-}
 variable "pf_cr_capacity" {
   type    = string
   default = "1Gbps" # 100Mbps