+++ internal/provider/resource_cloud_router.go
@@ -35,7 +35,7 @@ func resourceCloudRouter() *schema.Resource {
 			},
 			"asn": {
 				Type:        schema.TypeInt,
-				Required:    true,
+				Optional:    true,
 				Description: "The ASN of the cloud router.\n\n\tThis can be the PacketFabric public ASN 4556 (default) or a private ASN from 64512 - 65534.\n\n\tDefaults to 4556 if unspecified.",
 			},
 			"name": {
@@ -53,7 +53,7 @@ func resourceCloudRouter() *schema.Resource {
 			"regions": {
 				Type:        schema.TypeList,
 				Optional:    true,
-				Description: "The regions in which the Cloud Router connections will be located.\n\t\tUse `[\"US\"]` for North America and `[\"UK\"]` for EMEA. For transatlantic, use `[\"US\",\"UK\"]`.",
+				Description: "The regions in which the Cloud Router connections will be located.\n\t\tUse `[\"US\"]` for North America and `[\"UK\"]` for EMEA. For transatlantic, use `[\"US\",\"UK\"]`.\n\n\tDefaults to US if unspecified.",
 				Elem: &schema.Schema{
 					Type:         schema.TypeString,
 					ValidateFunc: validation.StringIsNotEmpty,